+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.4 (WIP)
+## 0.1.4 (2022-03-16)
 
 * Make fields work with the new `Alchemy.Map.Backed` and `Alchemy.Map.Enum` class
 * Add the `field_path_in_current_schema` property